- fixed warning type casting
git-svn-id: svn://rdo.rk9.bmstu.ru/var/svn/root@5425 23f926e6-80f1-49ca-85a0-60d4a27bb3ac

+++ rdo/packages/service/simulation/trunk/rdosimwin.cpp
@@ -734,7 +734,7 @@ void RDOThreadRunTime::writeResultsInfo()
 	
 	double delay = -1;
 
-	delay = timeStop - m_timeStart;
+	delay = (double)(timeStop - m_timeStart);
 	
 	if (delay != -1)
 	{
